CICD_model_training/requirements.txt
pandas
numpy
tensorflow>=2.12
keras>=2.12
kfp==2.7.0
google-cloud-aiplatform
